Reorganize some interface files.
