+++ Interface/Windows/DiskWindow.py
@@ -16,16 +16,16 @@
 # You should have received a copy of the GNU General Public License
 # along with ArchSetup.  If not, see <http://www.gnu.org/licenses/>.
 
-from SetupTools.Timezone import Timezone
-from Interface.SetupWindow import SetupWindow
-from Interface.SpacerWidget import SpacerWidget
-from Interface.TextWidget import TextWidget
-from Interface.ScrollWidget import ScrollWidget
-from Interface.RadioWidget import RadioWidget
+from SetupTools.Disks import Disks
+from Interface.Windows.SetupWindow import SetupWindow
+from Interface.Widgets.SpacerWidget import SpacerWidget
+from Interface.Widgets.TextWidget import TextWidget
+from Interface.Widgets.ScrollWidget import ScrollWidget
+from Interface.Widgets.RadioWidget import RadioWidget
 
 import gettext
 
-class TimesubzoneWindow(SetupWindow):
+class DiskWindow(SetupWindow):
     def __init__(self, callback, setupconfig):
         super().__init__()
 
@@ -33,25 +33,19 @@ class TimesubzoneWindow(SetupWindow):
         trans = gettext.translation("archsetup", "locale", fallback=True)
         trans.install()
 
-        # Init setup tools
-        self.timezone = Timezone()
         self.setupconfig = setupconfig
-
-        # Add widgets and setup callbacks...
-        self.addwidget(TextWidget(1, 1, _('Please select a subzone...'),  40))
-        self.radiowidget = RadioWidget(0, 0, 40, self.timezone.list_subzones(self.setupconfig.gettimezone()), self.event)
-        self.addwidget(ScrollWidget(3, 1, 40, 20, self.radiowidget, self.event))
+        self.addwidget(TextWidget(1, 1, _('Please select the disk, where ArchLinux should be installed'),  40))
+        disk = Disks()
+        items = disk.list_disks()
+        self.addwidget(ScrollWidget(4, 1, 40, 20, RadioWidget(0, 0, 40, items, self.event), self.event))
         self.addwidget(SpacerWidget(23, 1, 1))
         self.setnextcallback(callback, 'next')
         self.setprevcallback(callback, 'prev')
 
     def event(self, event, opt=''):
-        if event == 'show':
-            self.radiowidget.setlist(self.timezone.list_subzones(self.setupconfig.gettimezone()))
-            self.refresh()
-        elif event == 'refresh':
+        if event == 'refresh':
             self.refresh()
         elif event == 'selection':
-            self.setupconfig.settimesubzone(opt)
+            self.setupconfig.setdisk(opt)
         else:
             super().event(event)
