+++ Interface/Windows/FinishWindow.py
@@ -16,12 +16,12 @@
 # You should have received a copy of the GNU General Public License
 # along with ArchSetup.  If not, see <http://www.gnu.org/licenses/>.
 
-from Interface.SetupWindow import SetupWindow
-from Interface.TextWidget import TextWidget
+from Interface.Windows.SetupWindow import SetupWindow
+from Interface.Widgets.TextWidget import TextWidget
 
 import gettext
 
-class WelcomeWindow(SetupWindow):
+class FinishWindow(SetupWindow):
     def __init__(self, callback):
         super().__init__()
 
@@ -29,6 +29,7 @@ class WelcomeWindow(SetupWindow):
         trans = gettext.translation("archsetup", "locale", fallback=True)
         trans.install()
 
-        self.addwidget(TextWidget(1, 1, _('Welcome to the Archlinux installer!'),  40))
-        self.addwidget(TextWidget(3, 1, _('In order to navigate use the TAB key and press ENTER in order to activate the selected widget.'), 40))
+        self.addwidget(TextWidget(1, 1, _('All done!'),  40))
+        self.addwidget(TextWidget(3, 1, _('Archlinux is setup in your computer. The program will exit when you click in the next button.'), 40))
         self.setnextcallback(callback, 'next')
+        self.setprevcallback(callback, 'prev')
