+++ Interface/Windows/FontWindow.py
@@ -16,16 +16,16 @@
 # You should have received a copy of the GNU General Public License
 # along with ArchSetup.  If not, see <http://www.gnu.org/licenses/>.
 
-from SetupTools.Keyboard import Keyboard
-from Interface.SetupWindow import SetupWindow
-from Interface.SpacerWidget import SpacerWidget
-from Interface.TextWidget import TextWidget
-from Interface.ScrollWidget import ScrollWidget
-from Interface.RadioWidget import RadioWidget
+from SetupTools.Font import Font
+from Interface.Windows.SetupWindow import SetupWindow
+from Interface.Widgets.SpacerWidget import SpacerWidget
+from Interface.Widgets.TextWidget import TextWidget
+from Interface.Widgets.ScrollWidget import ScrollWidget
+from Interface.Widgets.RadioWidget import RadioWidget
 
 import gettext
 
-class KeyboardWindow(SetupWindow):
+class FontWindow(SetupWindow):
     def __init__(self, callback, setupconfig):
         super().__init__()
 
@@ -34,9 +34,9 @@ class KeyboardWindow(SetupWindow):
         trans.install()
 
         self.setupconfig = setupconfig
-        self.addwidget(TextWidget(1, 1, _('Please select a keyboard...'),  40))
-        keyboard = Keyboard()
-        items = keyboard.list_keyboard_layouts()
+        self.addwidget(TextWidget(1, 1, _('Please select a console font...'),  40))
+        font = Font()
+        items = font.list_console_fonts()
         self.addwidget(ScrollWidget(3, 1, 40, 20, RadioWidget(0, 0, 40, items, self.event), self.event))
         self.addwidget(SpacerWidget(23, 1, 1))
         self.setnextcallback(callback, 'next')
@@ -46,6 +46,6 @@ class KeyboardWindow(SetupWindow):
         if event == 'refresh':
             self.refresh()
         elif event == 'selection':
-            self.setupconfig.setkeyboard(opt)
+            self.setupconfig.setfont(opt)
         else:
             super().event(event)
