+++ Interface/Windows/TimesubzoneWindow.py
@@ -17,15 +17,15 @@
 # along with ArchSetup.  If not, see <http://www.gnu.org/licenses/>.
 
 from SetupTools.Timezone import Timezone
-from Interface.SetupWindow import SetupWindow
-from Interface.SpacerWidget import SpacerWidget
-from Interface.TextWidget import TextWidget
-from Interface.ScrollWidget import ScrollWidget
-from Interface.RadioWidget import RadioWidget
+from Interface.Windows.SetupWindow import SetupWindow
+from Interface.Widgets.SpacerWidget import SpacerWidget
+from Interface.Widgets.TextWidget import TextWidget
+from Interface.Widgets.ScrollWidget import ScrollWidget
+from Interface.Widgets.RadioWidget import RadioWidget
 
 import gettext
 
-class TimezoneWindow(SetupWindow):
+class TimesubzoneWindow(SetupWindow):
     def __init__(self, callback, setupconfig):
         super().__init__()
 
@@ -38,16 +38,20 @@ class TimezoneWindow(SetupWindow):
         self.setupconfig = setupconfig
 
         # Add widgets and setup callbacks...
-        self.addwidget(TextWidget(1, 1, _('Please select a time zone...'),  40))
-        self.addwidget(ScrollWidget(3, 1, 40, 20, RadioWidget(0, 0, 40, self.timezone.list_zones(), self.event), self.event))
+        self.addwidget(TextWidget(1, 1, _('Please select a subzone...'),  40))
+        self.radiowidget = RadioWidget(0, 0, 40, self.timezone.list_subzones(self.setupconfig.gettimezone()), self.event)
+        self.addwidget(ScrollWidget(3, 1, 40, 20, self.radiowidget, self.event))
         self.addwidget(SpacerWidget(23, 1, 1))
         self.setnextcallback(callback, 'next')
         self.setprevcallback(callback, 'prev')
 
     def event(self, event, opt=''):
-        if event == 'refresh':
+        if event == 'show':
+            self.radiowidget.setlist(self.timezone.list_subzones(self.setupconfig.gettimezone()))
+            self.refresh()
+        elif event == 'refresh':
             self.refresh()
         elif event == 'selection':
-            self.setupconfig.settimezone(opt)
+            self.setupconfig.settimesubzone(opt)
         else:
             super().event(event)
